--- dashWhat/src/components/ChatApp.jsx
@@ -1,15 +1,0 @@
-import React, { useState } from 'react';
-
-const ChatApp = () => {
-  const [selectedChat, setSelectedChat] = useState(null);
-
-  const handleChatClick = (remoteJid) => {
-    setSelectedChat(remoteJid);
-  };
-
-  return (
-    <div>nada</div>
-  )
-}
-
-export default ChatApp;
