+++ dashWhat/src/utils/orderStatus.js
@@ -1,11 +1,17 @@
 export const ORDER_STATUS = {
-  'No leido': 'bg-red-500 text-white',
   'Sin Confirmar': 'bg-red-500 text-white',
   'Confirmado': 'bg-green-500 text-white',
   'En proceso': 'bg-yellow-500 text-black',
   'Entregado': 'bg-blue-500 text-white',
   'Cancelado': 'bg-gray-800 text-white',
   // Añade más estados y colores según sea necesario
-};
+}
+export const CONVERSATION_STATUS = {
+  'No leido': 'bg-red-500 text-white',
+  'Leido': 'bg-green-500 text-white',
+  // Añade más estados y colores según sea necesario
+}
+
+
